hoàn thiện các module cơ bản

+++ BE/src/main/java/service/cinema/be/core/admin/quanlyve/repository/AdGiaVeChiTietRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import service.cinema.be.entity.GiaVeChiTiet;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 // ĐỔI Long -> String
@@ -19,7 +18,7 @@ public interface AdGiaVeChiTietRepository extends JpaRepository<GiaVeChiTiet, St
             "AND g.loaiGhe.id = :idLoaiGhe " +
             "AND g.khungGio.id = :idKhungGio " +
             "AND g.trangThai = 1")
-    Optional<BigDecimal> timGiaPhuThuTheoTieuChi(
+    Optional<Double> timGiaPhuThuTheoTieuChi(
             @Param("idLoaiKhach") String idLoaiKhach, // ĐỔI Long -> String
             @Param("idLoaiNgay") String idLoaiNgay,   // ĐỔI Long -> String
             @Param("idLoaiGhe") String idLoaiGhe,     // ĐỔI Long -> String
